PC-1432: rename PasswordAuth to Auth
