+++ README.md
@@ -196,7 +196,7 @@ Secrets must be configured in the ECS tasks, corresponding to the variables in `
 - `OsPlaces__Key`
 
 To prevent public access to DEV and UAT environments, we should also override the auth credentials:
-- `PasswordAuth__Password`
+- `Auth__Password`
 
 (These are not required for production)
