+++ WhlgPublicWebsite/appsettings.DEV.json
@@ -3,7 +3,7 @@
         "AppBaseUrl": "https://dev.check-eligibility-for-home-upgrade-grant.service.gov.uk/"
     },
 
-    "PasswordAuth": {
+    "Auth": {
         "Password": "warmhomes"
     },
