+++ WhlgPublicWebsite/appsettings.Staging.json
@@ -3,7 +3,7 @@
         "AppBaseUrl": "https://uat.check-eligibility-for-home-upgrade-grant.service.gov.uk/"
     },
 
-    "PasswordAuth": {
+    "Auth": {
         "Password": "warmhomes"
     },
